+++ main.go
@@ -1,11 +1,10 @@
 package main
 
 import (
+	"todoapp/auth"
 	"todoapp/tasks"
 	"todoapp/users"
-	"todoapp/auth"
 )
-
 
 func main() {
 
@@ -15,6 +14,6 @@ func main() {
 	r := auth.SetupRouter()
 	users.InitUsers(r)
 	tasks.InitTasks(r)
-	
-	r.Run(":8080") // Main File 
+
+	r.Run(":8080")
 }
