+++ tasks/tasksHandler.go
@@ -1,82 +1,27 @@
 package tasks
 
 import (
-	"fmt"
-	//"todoapp/users"
 	"github.com/gin-gonic/gin"
-	"net/http"
-	"strconv"
+	"todoapp/auth"
 )
 
-/////////////////////////////////////////////////////////////////////////////////////////////////
 
-// Listing All Tasks
-func GetTasks(c *gin.Context) {
+func InitTasks(r *gin.Engine) {
 	
-	var task0 []Task
-	DB.Find(&task0)
-	//fmt.Println(s)
-	c.JSON(http.StatusOK,task0)
-	
-}
+	r.GET("/tasks", auth.IsAuthorized(), GetTasks)
+	 
+	r.POST("/tasks", auth.IsAuthorized(), CreateTask)
 
-// Creating a New Task
-func CreateTask(c *gin.Context) {
-	
-	DB.AutoMigrate(&Task{})
-	//c.Set("UserID", Uid) 
-	//x = c.Get("UserID")
-	var task0 Task
-	r,_ := c.Cookie("Token")
-	if r == "" {
-		c.JSON(http.StatusForbidden,"Forbidden")
-		return
-	}
-	c.BindJSON(&task0)
-	y, _ := c.Get("client")
-	fmt.Println(y)
-	x, _ := strconv.Atoi(y.(string))
-	task0.Uid = x 
-	DB.Create(&task0)
-	fmt.Println(task0.ID)
-	c.JSON(http.StatusOK, "Task Added!")
-	
-}
+	r.PATCH("/tasks/:id", auth.IsAuthorized(), EditTask)
 
-// Task Deletion
-func DeleteTask(c *gin.Context) {
-	
-	var task0 Task
-	if e := DB.Where("id=?",c.Param("id")).First(&task0).Error; e != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"Error": "Task does not exist!"})
-		return 
-	}
+	r.DELETE("/tasks/:id", auth.IsAuthorized(), DeleteTask)
 
-	DB.Delete(&task0)
-	c.JSON(http.StatusOK, "Task Deleted Successfully!")
-}
+	r.DELETE("/all", auth.IsAuthorized(), DeleteAll)
 
-// Editing a Tasking Using ID as Key 
-func EditTask(c *gin.Context) {
+	// Reports
 
-	var task0 Task
-	if e := DB.Where("id=?",c.Param("id")).First(&task0).Error; e != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"Error": "Task does not exist!"})
-		return 
-	}
+	r.GET("/report/", auth.IsAuthorized(), CountTask)
 
-	
-	var input UpdateTask
-	c.BindJSON(&input)
-	fmt.Println("Com_status", input.Com_status)
-	DB.Model(&task0).Updates(Task{Title: input.Title, Description: input.Description, Com_status: input.Com_status})
-	c.JSON(http.StatusOK, "Task Modified Successfully!")
-}
+	r.GET("/report/mt", auth.IsAuthorized(), MissedTasks)
 
-// Clearing Whole DB
-func DeleteAll(c *gin.Context) {
-	var task0 []Task
-	DB.Find(&task0)
-	DB.Delete(&task0)
-	c.JSON(http.StatusOK, "All Entries Deleted!")
 }
