+++ users/usersServices.go
@@ -1,72 +1,71 @@
 package users
 
 import (
-	"fmt"
 	"net/http"
-	//"fmt"
-	"todoapp/auth"
 	"strconv"
+	"todoapp/auth"
 	"github.com/gin-gonic/gin"
+	"golang.org/x/crypto/bcrypt"
 )
-func VerifyEmail(x string) bool {
+func VerifyEmail(x string, c *gin.Context) bool {
 	var user User
-	if e := UDB.Where("Email=?", x).First(&user).Error; e == nil {
-		return false
+	out := UDB.Where("Email=?", x).First(&user)
+	flag := out.RowsAffected==0
+	if !flag {
+		c.JSON(http.StatusNotAcceptable, gin.H{"Error":"Email already registered"} )
 	}
-	return true
+	return flag
 }
 
 func CreateUser(c *gin.Context) {
-	
 	UDB.AutoMigrate(&User{})
-	var err error
 	var user User
 	c.BindJSON(&user)
+	ePassword, err := bcrypt.GenerateFromPassword([]byte(user.Password), bcrypt.DefaultCost)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, err.Error())
+	}
+	user.Password = string(ePassword)
 	
-
-	if VerifyEmail(user.Email) {
-		if err = UDB.Create(user).Error; err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"Error":"Failed to open db"} )
-					
-		} else {
-			c.JSON(http.StatusCreated ,"Account Registered")
-		}
-}
+	if VerifyEmail(user.Email, c) {
+		Count = Count + 1
+		user.IDU = Count
+		UDB.Create(user)
+		c.JSON(http.StatusCreated ,"Account Registered")
+	}
 }
 
 
 func Login(c *gin.Context) {
-	//fmt.Println("1--------------------------------------------------------------------")
-	var reqBody User
+	
+	var input User
 	var user User
-	c.BindJSON(&reqBody)
-	//fmt.Println("2--------------------------------------------------------------------")
-	//err := json.NewDecoder(r.Body).Decode(&reqBody)	fmt.Println("3--------------------------------------------------------------------")
-	
-	//fmt.Println("4--------------------------------------------------------------------")
-	//c.JSON(http.StatusOK, "Email: "+reqBody.Email+", Password: "+reqBody.Password)
-	//fmt.Println("5--------------------------------------------------------------------")
-	
-	out := UDB.Where("Email = ? AND Password = ?", reqBody.Email, reqBody.Password).Find(&user)
+	c.BindJSON(&input)
+
+	out := UDB.Where("Email = ?", input.Email).Find(&user)
 	if out.RowsAffected == 0 {
 		c.JSON(http.StatusNotFound, "Invalid Email or Password!")
 		return
 	}
 	
-	token, err := auth.CreateJWT(strconv.Itoa(user.ID), user.Password)
+	err := bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(input.Password))
+	if err == bcrypt.ErrMismatchedHashAndPassword {
+		c.JSON(http.StatusUnauthorized, "Invalid Email or Password!")
+		return
+	}
+
+	token,err := auth.CreateJWT(strconv.FormatUint(uint64(user.IDU), 10), user.Password)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, err.Error())
 		return
 	}
-	fmt.Println(token)
+
 	c.SetCookie("Token",token,1000000,"/","",false,false)
 
 	c.JSON(http.StatusOK, "User logged in successfully")
 
 }
-func Msg(c *gin.Context) {
-	c.JSON(200,"Secret Code")
-}
+
 func Logout(c *gin.Context) {
 	r,_ := c.Cookie("Token")
 	if r == "" {
@@ -77,19 +76,5 @@ func Logout(c *gin.Context) {
 	c.SetCookie("Token","",-1,"/","",false,false)
 
 	c.JSON(http.StatusOK,"Logged out successfully!")
-	//c.JSON(200,r)
+	
 }
-
-/*func logout(w http.ResponseWriter, r *http.Request) {
-	
-	_, err := auth.GetJWT(r)
-	if err != nil {
-		utils.JSONMsg(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-	
-	accessCookie := http.Cookie{Name: "accessCookie", Value: "", MaxAge: -1}
-	http.SetCookie(w, &accessCookie)
-	utils.JSONMsg(w, "User logged out successfully", http.StatusOK)
-
-}*/
